refactor: embed MtsTrackedSensor into MtsClimate and rename
