+++ tests/entities/select.py
@@ -1,5 +1,6 @@
 from homeassistant.components import select as haec
 
+from custom_components.meross_lan.climate import MtsClimate
 from custom_components.meross_lan.devices.diffuser import MLDiffuserSpray
 from custom_components.meross_lan.devices.spray import MLSpray
 from custom_components.meross_lan.devices.thermostat.mts300 import MtsSensorAssociation
@@ -10,7 +11,7 @@ from custom_components.meross_lan.devices.thermostat.mtsthermostat import (
     mn_t,
 )
 from custom_components.meross_lan.merossclient.protocol import const as mc
-from custom_components.meross_lan.select import MLSelect, MtsTrackedSensor
+from custom_components.meross_lan.select import MLSelect
 
 from tests.entities import EntityComponentTest
 
@@ -21,8 +22,8 @@ class EntityTest(EntityComponentTest):
 
     DIGEST_ENTITIES = {
         mc.KEY_THERMOSTAT: {
-            mc.KEY_MODE: [MtsTrackedSensor],
-            mc.KEY_MODEB: [MtsTrackedSensor],
+            mc.KEY_MODE: [MtsClimate.TrackSensorSelect],
+            mc.KEY_MODEB: [MtsClimate.TrackSensorSelect],
         },
         mc.KEY_SPRAY: [MLSpray],
         mc.KEY_DIFFUSER: {mc.KEY_SPRAY: [MLDiffuserSpray]},
@@ -31,12 +32,12 @@ class EntityTest(EntityComponentTest):
         mn.Appliance_Config_Sensor_Association.name: [MtsSensorAssociation],
         mn.Appliance_Control_TempUnit.name: [MtsTempUnit],
         mn_t.Appliance_Control_Thermostat_HoldAction.name: [MtsHoldAction],
-        mn_t.Appliance_Control_Thermostat_ModeC.name: [MtsTrackedSensor],
+        mn_t.Appliance_Control_Thermostat_ModeC.name: [MtsClimate.TrackSensorSelect],
     }
     HUB_SUBDEVICES_ENTITIES = {
-        mc.TYPE_MTS100: [MtsTrackedSensor],
-        mc.TYPE_MTS100V3: [MtsTrackedSensor],
-        mc.TYPE_MTS150: [MtsTrackedSensor],
+        mc.TYPE_MTS100: [MtsClimate.TrackSensorSelect],
+        mc.TYPE_MTS100V3: [MtsClimate.TrackSensorSelect],
+        mc.TYPE_MTS150: [MtsClimate.TrackSensorSelect],
     }
 
     async def async_test_enabled_callback(self, entity: MLSelect):
